Local DB script
The included script can be used to generate a copy of the database we
hope to use.  It's not a final version as Zorayr and I need to discuss
and make sure that this schema holds water to what we need to do with
it.

I've updated the .gitignore file to allow files of this extension.